🛠 refactor: 경로 단축을 위한 index.ts 파일 추가
