+++ src/auth/auth.module.ts
@@ -1,5 +1,4 @@
-import { User } from '../entities/user.entity';
-import { Profile } from '../entities/profile.entity';
+import { User, Profile, Password, Image, Oauth } from '../entities';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigService } from '@nestjs/config';
 import { Module } from '@nestjs/common';
@@ -8,23 +7,22 @@ import {
   LoginController,
   AuthLogoutController,
   AuthVerifyController,
+  GoogleLoginController,
 } from './controllers';
-import { SignupService, LoginService } from './services';
+import { SignupService, LoginService, GoogleLoginService } from './services';
 import { JwtModule, JwtService } from '@nestjs/jwt';
-import { PassportModule } from '@nestjs/passport';
-import { JwtStrategy } from './jwt.strategy';
+import { JwtStrategy } from './utils/jwt.strategy';
 import { CheckEmailDuplicateRepository } from './outbound-adapter/check-email-duplicate.repository';
 import { GetUserByEmailRepository } from './outbound-adapter/get-user-by-email.repository';
 import { CreateUserRepository } from './outbound-adapter/create-user.repository';
 import { GetUserByIdRepository } from './outbound-adapter/get-user-by-id.repository';
 import { CheckNicknameDuplicateRepository } from '../profile/outbound-adapter/check-nickname-duplicate.repository';
-import { Password } from '../entities/password.entity';
-import { Image } from '../entities/image.entity';
+import { GoogleStrategy } from './utils/google.strategy';
+import { CreateSocialUserRepository } from './outbound-adapter/create-social-user.repository';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([User, Profile, Password]),
-    PassportModule.register({ defaultStrategy: 'jwt' }),
+    TypeOrmModule.forFeature([User, Profile, Password, Image, Oauth]),
     JwtModule.registerAsync({
       inject: [ConfigService],
       useFactory: (configService: ConfigService) => ({
@@ -38,6 +36,7 @@ import { Image } from '../entities/image.entity';
     LoginController,
     AuthLogoutController,
     AuthVerifyController,
+    GoogleLoginController,
   ],
   providers: [
     SignupService,
@@ -49,6 +48,10 @@ import { Image } from '../entities/image.entity';
     JwtStrategy,
     JwtService,
     CheckNicknameDuplicateRepository,
+    GoogleLoginService,
+    GoogleStrategy,
+    GoogleLoginService,
+    CreateSocialUserRepository,
   ],
   exports: [],
 })
