+++ src/auth/controllers/index.ts
@@ -2,3 +2,4 @@ export * from './login.controller';
 export * from './signup.controller';
 export * from './logout.controller';
 export * from './verify.controller';
+export * from './google-login.controller';
