+++ src/auth/services/index.ts
@@ -1,2 +1,3 @@
 export * from './login.service';
 export * from './signup.service';
+export * from './google-login.service';
